# app/routers/qr.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from typing import Optional
from services.qr_service import PythonQRGenerator

router = APIRouter(prefix="/qr", tags=["qr-codes"])

class QRRequest(BaseModel):
    batch_id: int
    item_name: str
    donor_name: Optional[str] = "Royal Palace Banquet"
    quantity_kg: Optional[float] = 25.0
    plates_count: Optional[int] = 100
    safety_score: Optional[int] = 95

class QRVerifyRequest(BaseModel):
    token: str

@router.get("/generate/{batch_id}")
def generate_qr_by_batch_id(batch_id: int, item_name: Optional[str] = "Surplus Dish"):
    """
    Generate Python QR code image and metadata for a specific surplus batch ID.
    """
    try:
        data_url, payload = PythonQRGenerator.generate_qr_base64_png(
            batch_id=batch_id,
            item_name=item_name
        )
        svg_xml, _ = PythonQRGenerator.generate_qr_svg_str(
            batch_id=batch_id,
            item_name=item_name
        )
        return {
            "status": "success",
            "generator": "Python qrcode 8.2",
            "qr_image_data_url": data_url,
            "svg_xml": svg_xml,
            "payload": payload
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.post("/generate")
def generate_custom_qr(request: QRRequest):
    """
    Generate customized Python QR code payload and image.
    """
    try:
        data_url, payload = PythonQRGenerator.generate_qr_base64_png(
            batch_id=request.batch_id,
            item_name=request.item_name,
            donor_name=request.donor_name,
            quantity_kg=request.quantity_kg,
            plates_count=request.plates_count,
            safety_score=request.safety_score
        )
        svg_xml, _ = PythonQRGenerator.generate_qr_svg_str(
            batch_id=request.batch_id,
            item_name=request.item_name,
            donor_name=request.donor_name,
            quantity_kg=request.quantity_kg,
            plates_count=request.plates_count,
            safety_score=request.safety_score
        )
        return {
            "status": "success",
            "generator": "Python qrcode 8.2 Engine",
            "qr_image_data_url": data_url,
            "svg_xml": svg_xml,
            "payload": payload
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.post("/verify")
def verify_qr_token(request: QRVerifyRequest):
    """
    Verify scanned Python QR code token.
    """
    token = request.token.strip().upper()
    if not token.startswith("QR-FR-"):
        return {
            "verified": False,
            "message": "Invalid QR Code format. Must be generated by Food Rescue AI."
        }
    
    parts = token.split("-")
    batch_id = parts[2] if len(parts) >= 3 else "101"
    
    return {
        "verified": True,
        "message": f"QR Code token {token} verified authentic!",
        "batch_id": int(batch_id) if batch_id.isdigit() else 101,
        "status": "VERIFIED_AUTHENTIC",
        "verifier": "Python Security Gate"
    }